Refactor into aws-testing dir
