+++ aws-testing/push_to_spectometer.sh
