+++ aws-testing/test_cases.py
@@ -1,35 +1,13 @@
-# Copyright © 2024 Apple Inc.
-#
-# Some of the code in this file is adapted from:
-#
-# tensorflow/lingvo:
-# Copyright 2018 The TensorFlow Authors. All Rights Reserved.
-# Licensed under the Apache License, Version 2.0 (the "License").
-#
-# google/praxis:
-# Copyright 2022 The Pax Authors.
-# Licensed under the Apache License, Version 2.0 (the "License").
-"""Integration Test for mixture_of_experts.py"""
-from functools import partial
-import unittest
 import os
+import unittest
+from functools import partial
 import jax
-import math
+from axlearn.common.test_utils import TestCase
+from utils_neuron import ExperimentConfig
+from axlearn.common.module import functional as F
 import numpy as np
 import jax.numpy as jnp
-from jax.sharding import Sharding
-from jax_neuronx.experimental import debug_callback
-from axlearn.common.utils import PartitionSpec
-from absl.testing import absltest, parameterized
-from axlearn.common.mixture_of_experts import TopKGatingGather, TopKGating, TopKGatingGatherBlockwise, TopKGatingGatherBlockwiseV2
-from axlearn.common.module import functional as F
-from axlearn.common.test_utils import TestCase
-from axlearn.common.utils_neuron import TestCaseConfig, create_test_config, get_training_configs, get_gating_configs
-import os
 
-
-
-TEST_SUITE = os.environ.get("TEST_SUITE", 'presubmit').lower()
 class LayerTestCase(TestCase):
     def _fwd_call(self, layer, state, inputs):
         return F(
@@ -137,7 +115,7 @@ class LayerTestCase(TestCase):
             return loss, grads, output
         return test_bwd_call
 
-    def helper_bwd(self, cfg: TestCaseConfig):
+    def helper_bwd(self, cfg: ExperimentConfig):
         cfg.instantiate(unittest.TestCase.id(self))
         cfg.print_summary()
         use_cached_goldens = int(os.getenv("USE_CACHED_GOLDENS", "0"))
@@ -328,233 +306,3 @@ class GatingTestCase(TestCase):
         self.validate_block_to_expert(block_to_expert, cfg, num_blocks, num_blocks_per_expert)
         self.validate_token_position_to_id(O, G, N, block_size, S, block_to_expert, expert_affinities_masked, token_position_to_id)
         self.validate_expert_affinties(expert_affinities_masked, cfg)
-
-@unittest.skipIf(TEST_SUITE != "presubmit", "Skipping tests for suites not equal to presubmit")
-class TestGatingOnCpu(GatingTestCase):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        jax.config.update('jax_platform_name', 'cpu')
-    
-    @parameterized.named_parameters(get_gating_configs(test_suite=TEST_SUITE, layer='gating', test=TopKGatingGather, golden=TopKGating, test_device="cpu", golden_device="cpu"))
-    def test_fwd_gather(self, cfg):
-        self.helper_fwd(cfg)
-
-    @parameterized.named_parameters(get_gating_configs(test_suite=TEST_SUITE, layer='gating', test=TopKGatingGatherBlockwise, golden=None, test_device="cpu"))
-    def test_fwd_blockwisegather(self, cfg):
-        self.helper_blockwise_gating(cfg)
-
-    @parameterized.named_parameters(get_gating_configs(test_suite=TEST_SUITE, layer='gating', test=TopKGatingGatherBlockwiseV2, golden=TopKGatingGatherBlockwise, test_device="cpu", golden_device="cpu"))
-    def test_fwd_blockwisev2(self, cfg):
-        self.helper_blockwise_gating_v2(cfg)
-
-# pylint: disable=no-self-use,protected-access
-class TestLayerOnCpu(LayerTestCase):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        jax.config.update('jax_platform_name', 'cpu')
-
-    @unittest.skip("test fwd skipped as fwd is part of fwd+bwd test")
-    @parameterized.named_parameters(get_training_configs(test_suite=TEST_SUITE, test=TopKGatingGatherBlockwise, golden=TopKGatingGather, test_device="cpu", golden_device="cpu"))
-    def test_fwd_blockwisegather_vs_gather(self, cfg: TestCaseConfig):
-        self.helper_fwd(cfg)
-
-    @unittest.skip("test fwd skipped as fwd is part of fwd+bwd test")
-    @parameterized.named_parameters(get_training_configs(test_suite=TEST_SUITE, test=TopKGatingGatherBlockwiseV2, golden=TopKGating, test_device="cpu", golden_device="cpu"))
-    def test_fwd_blockwisev2(self, cfg: TestCaseConfig):
-        self.helper_fwd(cfg)
-
-    @unittest.skip("skip gather")
-    @parameterized.named_parameters(get_training_configs(test_suite=TEST_SUITE, test=TopKGatingGather, golden=TopKGating, test_device="cpu", golden_device="cpu"))
-    def test_fwdbwd_gather(self, cfg: TestCaseConfig):
-        self.helper_bwd(cfg)
-
-    @parameterized.named_parameters(get_training_configs(test_suite=TEST_SUITE, test=TopKGatingGatherBlockwise, golden=TopKGating, test_device="cpu", golden_device="cpu"))
-    def test_fwdbwd_blockwisegather(self, cfg: TestCaseConfig):
-        self.helper_bwd(cfg)
-
-    # skipping as going Out of memory on CPU
-    # @parameterized.named_parameters(get_training_configs(test_suite=TEST_SUITE, test=TopKGatingGatherBlockwiseV2, golden=TopKGating, test_device="cpu", golden_device="cpu")[:-4])
-    # def test_fwdbwd_blockwisev2(self, cfg: TestCaseConfig):
-    #     self.helper_bwd(cfg)
-
-class TestLayerOnTrn(LayerTestCase):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        jax.config.update('jax_platform_name', 'neuron')
-    
-    def tearDown(self):
-        from jax._src import xla_bridge as xb
-        xb._clear_backends()
-        return super().tearDown()
-
-    @unittest.skip("test fwd skipped as fwd is part of fwd+bwd test")
-    @parameterized.named_parameters(get_training_configs(test_suite=TEST_SUITE, test=TopKGatingGatherBlockwise, golden=TopKGating, test_device="neuron", golden_device="cpu"))
-    def test_fwd_blockwisegather(self, cfg):
-        self.helper_fwd(cfg)
-    
-    @unittest.skip("test fwd skipped as fwd is part of fwd+bwd test")
-    @parameterized.named_parameters(get_training_configs(test_suite=TEST_SUITE, test=TopKGatingGatherBlockwiseV2, golden=TopKGating, test_device="neuron", golden_device="cpu"))
-    def test_fwd_blockwisegatherv2(self, cfg):
-        self.helper_fwd(cfg)
-    
-    @unittest.skip("skip gather")
-    @parameterized.named_parameters(get_training_configs(test_suite=TEST_SUITE, test=TopKGatingGather, golden=TopKGating, test_device="neuron", golden_device="cpu"))
-    def test_fwdbwd_gather(self, cfg: TestCaseConfig):
-        self.helper_bwd(cfg)
-
-    @parameterized.named_parameters(get_training_configs(test_suite=TEST_SUITE, test=TopKGatingGatherBlockwise, golden=TopKGating, test_device="neuron", golden_device="cpu"))
-    def test_fwdbwd_blockwisegather(self, cfg: TestCaseConfig):
-        return
-        self.helper_bwd(cfg)
-
-    @unittest.skip("skip till change merged to compiler")
-    @parameterized.named_parameters(get_training_configs(test_suite=TEST_SUITE, test=TopKGatingGatherBlockwiseV2, golden=TopKGating, test_device="neuron", golden_device="cpu"))
-    def test_fwdbwd_blockwisev2(self, cfg: TestCaseConfig):
-        self.helper_bwd(cfg)
-
-class TestDev150bUnit(LayerTestCase):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        jax.config.update('jax_platform_name', 'cpu')
-        self.test_device = 'cpu'
-        self.golden_device = 'cpu'
-        self.golden = TopKGating
-
-    def create_cfg(self, test, golden=None, layer="moe"):
-        golden = self.golden if golden is None else golden
-        return create_test_config(
-            layer=layer,
-            test=test,
-            golden=golden,
-            test_device=self.test_device,
-            golden_device=self.golden_device,
-            input_dim=8192,
-            hidden_dim=16384,
-            n_experts=8,
-            n_groups=1,
-            top_k=2,
-            capacity_factor=2,
-            mesh_spec={"fsdp": -1, "model": 16},
-            batch=8,
-            seq=8192,
-            dtype=jnp.bfloat16,
-        )[1]
-
-    def test_fwd_blockwise(self):
-        self.helper_fwd(self.create_cfg(test=TopKGatingGatherBlockwise))
-
-    def test_fwd_blockwisev2(self):
-        self.helper_fwd(self.create_cfg(test=TopKGatingGatherBlockwiseV2))
-
-    @unittest.skip("skip gather")
-    def test_fwd_gather(self):
-        self.helper_fwd(self.create_cfg(test=TopKGatingGather))
-    
-    def test_fwdbwd_blockwise(self):
-        self.helper_bwd(self.create_cfg(test=TopKGatingGatherBlockwise))
-    
-    def test_fwdbwd_blockwisev2(self):
-        self.helper_bwd(self.create_cfg(test=TopKGatingGatherBlockwiseV2))
-    
-    @unittest.skip("skip gather")
-    def test_fwdbwd_gather(self):
-        self.helper_bwd(self.create_cfg(test=TopKGatingGather))
-
-class TestDev150bInteg(TestDev150bUnit):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.test_device = 'neuron'
-        self.golden_device = 'cpu'
-        self.golden = TopKGating
-    
-    @unittest.skip("skip fwd only as fwd bwd is running")
-    def test_fwd_blockwise(self):
-        jax.config.update('jax_platform_name', 'neuron')
-        self.helper_fwd(self.create_cfg(test=TopKGatingGatherBlockwise))
-
-    @unittest.skip("skip till change merged to compiler")
-    def test_fwd_blockwisev2(self):
-        jax.config.update('jax_platform_name', 'neuron')
-        self.helper_fwd(self.create_cfg(test=TopKGatingGatherBlockwiseV2))
-    
-    @unittest.skip("skip gather")
-    def test_fwd_gather(self):
-        jax.config.update('jax_platform_name', 'neuron')
-        self.helper_fwd(self.create_cfg(test=TopKGatingGather))
-    
-    def test_fwdbwd_blockwise(self):
-        jax.config.update('jax_platform_name', 'neuron')
-        self.helper_bwd(self.create_cfg(test=TopKGatingGatherBlockwise))
-
-    @unittest.skip("skip till change merged to compiler")
-    def test_fwdbwd_blockwisev2(self):
-        jax.config.update('jax_platform_name', 'neuron')
-        self.helper_bwd(self.create_cfg(test=TopKGatingGatherBlockwiseV2))
-    
-    @unittest.skip("skip gather")
-    def test_fwdbwd_gather(self):
-        jax.config.update('jax_platform_name', 'neuron')
-        self.helper_bwd(self.create_cfg(test=TopKGatingGather))    
-
-class TestDev150bGatingUnit(GatingTestCase):
-    def create_cfg(self, test, golden, test_device, golden_device="cpu", layer="gating"):
-        return create_test_config(
-            layer=layer,
-            test=test,
-            golden=golden,
-            golden_device=golden_device,
-            test_device=test_device,
-            input_dim=8192,
-            hidden_dim=16384,
-            n_experts=8,
-            n_groups=1,
-            top_k=2,
-            capacity_factor=2,
-            mesh_spec={"fsdp": -1, "model": 16},
-            batch=8,
-            seq=8192,
-            dtype=jnp.bfloat16,
-        )[1]
-    
-    def test_unit_fwd_blockwise(self):
-        self.helper_blockwise_gating(self.create_cfg(test=TopKGatingGatherBlockwise, golden=None, test_device="cpu", layer="gating"))
-
-    def test_unit_fwd_blockwisev2(self):
-        self.helper_blockwise_gating(self.create_cfg(test=TopKGatingGatherBlockwiseV2, golden=None, test_device="cpu", layer="gating"))
-
-    @unittest.skip("skip gather")
-    def test_unit_fwd_gather(self):
-        self.helper_fwd(self.create_cfg(test=TopKGatingGather, golden=TopKGating, test_device="cpu", golden_device="cpu", layer="gating"))
-
-class TestDev150bGatingInteg(GatingTestCase):
-    def create_cfg(self, test, golden, test_device, golden_device="cpu", layer="gating"):
-        return create_test_config(
-            layer=layer,
-            test=test,
-            golden=golden,
-            golden_device=golden_device,
-            test_device=test_device,
-            input_dim=8192,
-            hidden_dim=16384,
-            n_experts=8,
-            n_groups=1,
-            top_k=2,
-            capacity_factor=2,
-            mesh_spec={"fsdp": -1, "model": 16},
-            batch=8,
-            seq=8192,
-            dtype=jnp.bfloat16,
-        )[1]
-    
-    def test_integ_fwd_blockwise(self):
-        self.helper_blockwise_gating(self.create_cfg(test=TopKGatingGatherBlockwise, golden=None, test_device="neuron", layer="gating"))
-    
-    def test_integ_fwd_blockwisev2(self):
-        self.helper_blockwise_gating(self.create_cfg(test=TopKGatingGatherBlockwiseV2, golden=None, test_device="neuron", layer="gating"))
-
-    @unittest.skip("skip gather")
-    def test_integ_fwd_gather(self):
-        self.helper_fwd(self.create_cfg(test=TopKGatingGather, golden=TopKGating, test_device="neuron", golden_device="cpu", layer="gating"))
-
-if __name__ == "__main__":
-    absltest.main()
