+++ aws-testing/unit_test.sh
@@ -12,8 +12,10 @@ export TEST_LOG_DIR=${3:-"test_artifacts/shell"}
 export GOLDENS_DIR=${4:-"test_goldens"}
 export JAX_COMPILATION_CACHE_DIR=${5:-"test_artifacts/shell_jax_cc"}
 
-
-mkdir -p "${JAX_COMPILATION_CACHE_DIR}"
+# if defined
+if [ -n "$1" ]; then
+    mkdir -p "${JAX_COMPILATION_CACHE_DIR}"
+fi
 mkdir -p "${GOLDENS_DIR}"
 
 export TEST_ARTIFACTS_PATH=$TEST_LOG_DIR/$TEST_SUITE/artifacts
@@ -75,45 +77,44 @@ export NEURON_CC_FLAGS="${NEURON_CC_FLAGS} --internal-hlo2tensorizer-options='--
 export NEURON_CC_FLAGS="${NEURON_CC_FLAGS} --auto-cast=none" # --hbm-scratchpad-page-size=1024"
 export TF_CPP_MIN_LOG_LEVEL=3
 
-set -ex
-
 if [ "$1" = "unit" ]; then
     export JAX_PLATFORMS=cpu
-    pytest -rsA --tb=short --junitxml=$TEST_LOG_DIR/$TEST_SUITE/unit.xml axlearn/common/mixture_of_experts_neuron_test.py::TestLayerOnCpu
+    pytest -rsA --tb=short --junitxml=$TEST_LOG_DIR/$TEST_SUITE/unit.xml aws-testing/moe_layer_unit_test.py
 elif [ "$1" = "integ" ]; then
-    # if 150b or presubmit, break into two parts
     # breaking them up as we seem to leak memory across tests
-    if [ "$2" = "12b" ] || [ "$2" = "50b" ]; then
+    if [ "$2" = "12b" ] || [ "$2" = "50b" ] || [ "$2" = "switch-xxl" ] || [ "$2" = "llama4-maverick" ] || [ "$2" = "qwen3-235b" ]; then
         export TEST_SUITE_PARTS=1
     else
         export TEST_SUITE_PARTS=10
     fi
+    echo "Splitting tests into $TEST_SUITE_PARTS parts"
     status=0
     set +e
     for ((part=0;part<TEST_SUITE_PARTS;part++)); do
-        TEST_SUITE_PART=$part pytest -rsA --tb=short --junitxml=$TEST_LOG_DIR/$TEST_SUITE/integ_$part.xml axlearn/common/mixture_of_experts_neuron_test.py -k "TestLayerOnTrn"
+        # --collect-only -q
+        # use above if you only want to see the tests that will be run
+        set -x
+        TEST_SUITE_PART=$part pytest -rsA --tb=short --junitxml=$TEST_LOG_DIR/$TEST_SUITE/integ_$part.xml aws-testing/moe_layer_integ_test.py -k "TestLayerOnTrn"
         status_part=$?
         status=$((status + status_part))
+        set +x
     done
     if [ $status -ne 0 ]; then
         exit 1
     fi
 elif [ "$1" = "150bdev" ]; then
     export TEST_SUITE="150b"
-    pytest -rsA --tb=short --junitxml=$TEST_LOG_DIR/$TEST_SUITE/150bdev_integ.xml axlearn/common/mixture_of_experts_neuron_test.py -k "TestDev150bInteg or TestDev150bGatingInteg"
+    pytest -rsA --tb=short --junitxml=$TEST_LOG_DIR/$TEST_SUITE/150bdev_layer_integ.xml aws-testing/moe_layer_integ_test.py -k "TestDev150bInteg"
+    pytest -rsA --tb=short --junitxml=$TEST_LOG_DIR/$TEST_SUITE/150bdev_gating_integ.xml aws-testing/gating_test.py -k "TestDev150bGatingInteg"
     export JAX_PLATFORMS=cpu
-    pytest -rsA --tb=short --junitxml=$TEST_LOG_DIR/$TEST_SUITE/150bdev_unit.xml axlearn/common/mixture_of_experts_neuron_test.py -k "TestDev150bUnit or TestDev150bGatingUnit"
+    pytest -rsA --tb=short --junitxml=$TEST_LOG_DIR/$TEST_SUITE/150bdev_layer_unit.xml aws-testing/moe_layer_unit_test.py -k "TestDev150bUnit"
+    pytest -rsA --tb=short --junitxml=$TEST_LOG_DIR/$TEST_SUITE/150bdev_gating_unit.xml aws-testing/gating_test.py -k "TestGatingOnCpu or TestDev150bGatingUnit"
 elif [ "$1" = "dev" ]; then
-    pytest -rsA -v axlearn/common/mixture_of_experts_neuron_test.py -k "TestLayerOnTrn and test_fwdbwd_blockwisegather_MoE_i8192_h20480_e16_topk2_g1_ec2_blocksize512_b1_s8192_meshfsdp-1tp64_bf16"
-elif [ "$1" = "150b_gather" ]; then
-    pytest -rsA --tb=short axlearn/common/mixture_of_experts_neuron_test.py -k 'TestDev150bUnit and test_fwd_gather_vs_einsum or TestDev150bUnit and test_fwdbwd_gather_vs_einsum or TestDev150bInteg and test_fwd_gather_vs_einsum or TestDev150bInteg and test_fwdbwd_gather_vs_einsum'
-elif [ "$1" = "150b_blockwise" ]; then
-    pytest -rsA --tb=short axlearn/common/mixture_of_experts_neuron_test.py -k 'TestDev150bUnit and test_fwd_blockwise_vs_einsum or TestDev150bUnit and test_fwdbwd_blockwise_vs_einsum or TestDev150bInteg and test_fwd_blockwise_vs_einsum or TestDev150bInteg and test_fwdbwd_blockwise_vs_einsum'
-    pytest -rsA --tb=short axlearn/common/mixture_of_experts_neuron_test.py -k 'TestDev150bUnit and test_fwd_blockwisev2_vs_einsum or TestDev150bUnit and test_fwdbwd_blockwisev2_vs_einsum or TestDev150bInteg and test_fwd_blockwisev2_vs_einsum or TestDev150bInteg and test_fwdbwd_blockwisev2_vs_einsum'
+    pytest -rsA -v aws-testing/moe_layer_integ_test.py -k "TestLayerOnTrn and test_fwdbwd_blockwisegather_MoE_i8192_h20480_e16_topk2_g1_ec2_blocksize512_b1_s8192_meshfsdp-1tp64_bf16"
 elif [ "$1" = "150b_blockwise_cpu" ]; then
-    pytest -rsA --tb=short axlearn/common/mixture_of_experts_neuron_test.py -k 'TestDev150bUnit and test_fwd_blockwise_vs_einsum or TestDev150bUnit and test_fwdbwd_blockwise_vs_einsum'
-    pytest -rsA --tb=short axlearn/common/mixture_of_experts_neuron_test.py -k 'TestDev150bUnit and test_fwd_blockwisev2_vs_einsum or TestDev150bUnit and test_fwdbwd_blockwisev2_vs_einsum'
+    pytest -rsA --tb=short aws-testing/moe_layer_unit_test.py -k 'TestDev150bUnit and test_fwd_blockwise_vs_einsum or TestDev150bUnit and test_fwdbwd_blockwise_vs_einsum'
+    pytest -rsA --tb=short aws-testing/moe_layer_unit_test.py -k 'TestDev150bUnit and test_fwd_blockwisev2_vs_einsum or TestDev150bUnit and test_fwdbwd_blockwisev2_vs_einsum'
 elif [ "$1" = "150b_blockwise_neuron" ]; then
-    pytest -rsA --tb=short axlearn/common/mixture_of_experts_neuron_test.py -k 'TestDev150bInteg and test_fwd_blockwise_vs_einsum or TestDev150bInteg and test_fwdbwd_blockwise_vs_einsum'
-    pytest -rsA --tb=short axlearn/common/mixture_of_experts_neuron_test.py -k 'TestDev150bInteg and test_fwd_blockwisev2_vs_einsum or TestDev150bInteg and test_fwdbwd_blockwisev2_vs_einsum'
+    pytest -rsA --tb=short aws-testing/moe_layer_integ_test.py -k 'TestDev150bInteg and test_fwd_blockwise_vs_einsum or TestDev150bInteg and test_fwdbwd_blockwise_vs_einsum'
+    pytest -rsA --tb=short aws-testing/moe_layer_integ_test.py -k 'TestDev150bInteg and test_fwd_blockwisev2_vs_einsum or TestDev150bInteg and test_fwdbwd_blockwisev2_vs_einsum'
 fi
